# optimize the low cardinality column merge process