I hereby agree to the terms of the CLA available at: https://yandex.ru/legal/cla/?lang=en

Changelog category (leave one):
- Performance Improvement

Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):

Optimized the low cardinality column merge process for the vertical merge algorithm. 

Detailed description / Documentation draft:
The improvement focus on the low cardinality column merge process, and implemented a new way to gather the low cardinality column during vertical merge.
Before the improvement, during vertical merge, each block would be gathered from several parts, and each insert would compute the dictionary, and each assembled block would cost many times dictionary compute / index cut etc.
And this PR improve that by build a shared dictionary among the merged parts. The index of the merged parts will be transformed based on the shared dictionary, and then during the merge process only insert the index to the result column.

A little performance test:

```
CREATE TABLE default.LCDict
(
    `d2` UInt32,
    `d1` UInt32,
    `uintlc` LowCardinality(UInt64),
    `fltlc` LowCardinality(Float64),
    `strlc` LowCardinality(String)
)
ENGINE = MergeTree
PARTITION BY d2
ORDER BY (d2, d1)
SETTINGS index_granularity = 8192, vertical_merge_algorithm_min_rows_to_activate = 100000, vertical_merge_algorithm_min_columns_to_activate = 2, min_bytes_for_wide_part = 10000


INSERT INTO LCDict  SELECT intDiv(number, 5000000) AS d2,number AS d1,  (rand64() % 7000 +1)*10000 AS uint,    uint * pi() as flt, ['one', 'two', 'three', 'four', 'five', 'six', 'seven', 'eight', 'nine', 'ten'][rand()%10+1] AS str FROM numbers(50000000) order by rand()

**Before**:
2021.07.19 17:52:25.304170 [ 2388944 ] {f565d0cc-a597-453b-8fe5-f2a7d8c0f9cd} <Debug> ColumnGathererStream: Gathered column uintlc (2.859778113188539 bytes/elem.) in 0.344000575 sec., 3231791.1096514883 rows/sec., 8.81 MiB/sec.
2021.07.19 17:52:25.664465 [ 2388944 ] {f565d0cc-a597-453b-8fe5-f2a7d8c0f9cd} <Debug> ColumnGathererStream: Gathered column fltlc (2.859778113188539 bytes/elem.) in 0.356000595 sec., 3122854.3311844748 rows/sec., 8.52 MiB/sec.
2021.07.19 17:52:25.956497 [ 2388944 ] {f565d0cc-a597-453b-8fe5-f2a7d8c0f9cd} <Debug> ColumnGathererStream: Gathered column strlc (1.0022343393857185 bytes/elem.) in 0.288000481 sec., 3860194.941827198 rows/sec., 3.69 MiB/sec.
2021.07.19 17:52:25.957880 [ 2388944 ] {f565d0cc-a597-453b-8fe5-f2a7d8c0f9cd} <Debug> default.LCDict (7bfa27cc-9106-4f5a-bbfa-27cc91066f5a) (MergerMutator): Merge sorted 1111738 rows, containing 5 columns (2 merged, 3 gathered) in 1.1057698 sec., 1005397.3259172026 rows/sec., 26.01 MiB/sec.

**After the improvement**:
2021.07.19 17:54:02.875385 [ 2391020 ] {d59e2e5b-3f9a-474a-b9c1-d7230599766e} <Debug> ColumnGathererStream: Gathered column uintlc (2.1006459287798154 bytes/elem.) in 0.120000201 sec., 9274742.798139146 rows/sec., 18.58 MiB/sec.
2021.07.19 17:54:03.000718 [ 2391020 ] {d59e2e5b-3f9a-474a-b9c1-d7230599766e} <Debug> ColumnGathererStream: Gathered column fltlc (2.1006459287798154 bytes/elem.) in 0.120000201 sec., 9274742.798139146 rows/sec., 18.58 MiB/sec.
2021.07.19 17:54:03.036786 [ 2391020 ] {d59e2e5b-3f9a-474a-b9c1-d7230599766e} <Debug> ColumnGathererStream: Gathered column strlc (1.0002479848980792 bytes/elem.) in 0.032000054 sec., 34780285.058268964 rows/sec., 33.18 MiB/sec.
2021.07.19 17:54:03.038090 [ 2391020 ] {d59e2e5b-3f9a-474a-b9c1-d7230599766e} <Debug> default.LCDict (7bfa27cc-9106-4f5a-bbfa-27cc91066f5a) (MergerMutator): Merge sorted 1112971 rows, containing 5 columns (2 merged, 3 gathered) in 0.392421579 sec., 2836161.565926526 rows/sec., 73.32 MiB/sec.

```

The test result shows more efficient after the improvement, especially with many low cardinality columns, and the gathered block assembled scrappy from many parts.


Information about CI checks: https://clickhouse.tech/docs/en/development/continuous-integration/