@compasses 2021-07-19
@KochetovNicolai Hi Nikolai, this is the PR. 
BTW, for the low cardinality to be massively available, I think we need make the low cardinality column fall back to normal column when the dictionary size big enough.  Could you shed some more light about this? I found some of you presentation mentioned that,  and  I can continue working on it.