@compasses 2021-07-20
> @compasses Hi. Thank you for PR!
> 
> > we need make the low cardinality column fall back to normal column when the dictionary size big enough
> 
> Yes, it would be good to support.
> However, the proper support, from my point of view, requires a major refactoring.
> 
> Ideally, `LowCardinality` should not be a separate type. Type should define a set of values it can store. If so, then `LowCardinality(X)` and `X` define identical sets and should not be different. (`Nullable(X)` and `X` are different types still, cause the first one can store `Null`).
> `LowCardinality` is more like a codec, or data representation (as well as we can use `ColumnConst`, `ColumnSparse` or `ColumnVector` to represent values of the same type). This way, all the code should support any column representation, having only information about the type (or it could require specific representation via conversion).
> This is a nice goal to achieve, because it would be possible to easily add any other representation.
> 
> We can probably use a different approach. Like, store optional fill column (like, ColumnString) inside `ColumnLowCardinality`. So that it will work like if index is an identity, and dictionary is a full column (but not with uniq values anymore). It will be probably easier to implement, but it seem to me like a more complicated solution in afterwards.
> 
> In both approaches, we should also decide how we would serialize binary data.
> In the first approach, we need to support different representation (for granule?) externally. So that, we need to change, e.g., format for `.mrk` file in `MergeTree`. Again, it is more difficult, but more generic and powerful implementation.
> In the second approach, it is possible to update only `LowCardinality` serialization.