@compasses 2021-07-21
@KochetovNicolai Hi. Thanks for your great idea about the fallback solution.

About the fallback solution, I may prefer the second one.
For the first one, as you said, it's more difficult and also needs major refactoring. 
But for the second one, I have some concerns.
1. If we store a full column instead of a unique column in the low cardinality column, it seems many methods will be polluted. Like I need to know if the low cardinality column is in full state or unique state. So there may be full of if/else checks.
2. If in a full state, the index column is redundant too. It's better to bypass it in all the class member function.

So I think the second solution could be implemented except it may make the low cardinality column source code looks ugly with many if/else checks.