@compasses 2021-07-26
@KochetovNicolai 
> That's one of the reasons why it is not implemented yet.
> Many if/esle check may be expensive (it need to be checked).
Hi, 
I think I will give it a try :), maybe just for internal use only.